// parallel-web/src/types/CreateProjectRequest.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { RepoConfig } from "./RepoConfig";
import type { SshKeyConfig } from "./SshKeyConfig";

export type CreateProjectRequest = { name: string, repos: Array<RepoConfig>, ssh_keys: Array<SshKeyConfig>, parent_id?: string, };