// parallel-web/src/types/Project.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { RepoConfig } from "./RepoConfig";
import type { SshKeyConfig } from "./SshKeyConfig";

export type Project = { id: string, name: string, repos: Array<RepoConfig>, ssh_keys: Array<SshKeyConfig>, created_at: string, updated_at: string, parent_id?: string, };